unzipped_files/SolanaMemoTrader/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.18",
    "alembic>=1.15.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "qrcode>=8.2",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.40",
    "telegram>=0.0.1",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
]
